pyproject.toml
[tool.poetry]
name = "vcounter"
version = "2021.0.1"
description = "Applications for calculating statistics of visits"
authors = ["KirillShoniya"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.9"
PyYAML = "^5.4.1"
Flask = "^2.0.1"
redis = "^3.5.3"
python-dotenv = "^0.19.0"
setuptools = "^57.4.0"
gunicorn = "^20.1.0"
loguru = "^0.5.3"

[tool.poetry.dev-dependencies]
pytest = "^6.2.4"
coverage = "^5.5"
mypy = "^0.910"
locust = "^2.0.0"
types-PyYAML = "^5.4.3"
types-redis = "^3.5.4"

[build-system]
requires = [
    "poetry-core>=1.0.0",
    "setuptools>=30.3.0,<50",
    "wheel",
    "pytest-runner",
    "setuptools_scm>=3.3.1",
]
build-backend = "poetry.core.masonry.api"
